refactor(amazonq): refactor base lsp config
