+++ packages/core/src/amazonq/index.ts
@@ -37,7 +37,6 @@ export { ExtensionMessage } from '../amazonq/webview/ui/commands'
 export { CodeReference } from '../codewhispererChat/view/connector/connector'
 export { extractAuthFollowUp } from './util/authUtils'
 export { Messenger } from './commons/connector/baseMessenger'
-export * from './lsp/config'
 export * as secondaryAuth from '../auth/secondaryAuth'
 export * as authConnection from '../auth/connection'
 export * as featureConfig from './webview/generators/featureConfig'
